+++ core/src/main/kotlin/com/gitlab/kordlib/core/behavior/WebhookBehavior.kt
@@ -2,8 +2,10 @@ package com.gitlab.kordlib.core.behavior
 
 import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.core.Kord
+import com.gitlab.kordlib.core.cache.data.MessageData
 import com.gitlab.kordlib.core.cache.data.WebhookData
 import com.gitlab.kordlib.core.entity.Entity
+import com.gitlab.kordlib.core.entity.Message
 import com.gitlab.kordlib.core.entity.Webhook
 import com.gitlab.kordlib.rest.builder.webhook.ExecuteWebhookBuilder
 import com.gitlab.kordlib.rest.builder.webhook.WebhookModifyBuilder
@@ -35,7 +37,7 @@ interface WebhookBehavior : Entity {
 
             override fun hashCode(): Int = Objects.hash(id)
 
-            override fun equals(other: Any?): Boolean = when(other) {
+            override fun equals(other: Any?): Boolean = when (other) {
                 is WebhookBehavior -> other.id == id
                 else -> false
             }
@@ -73,6 +75,25 @@ suspend inline fun WebhookBehavior.edit(token: String, builder: WebhookModifyBui
 /**
  * Requests to execute this webhook.
  */
-suspend inline fun WebhookBehavior.execute(token: String, builder: ExecuteWebhookBuilder.() -> Unit) {
-    kord.rest.webhook.executeWebhook(token = token, webhookId = id.value, wait = true, builder = builder)
+suspend inline fun WebhookBehavior.execute(token: String, builder: ExecuteWebhookBuilder.() -> Unit): Message {
+    val response = kord.rest.webhook.executeWebhook(
+            token = token,
+            webhookId = id.value,
+            wait = true,
+            builder = builder
+    )!!
+
+    val data = MessageData.from(response)
+
+    return Message(data, kord)
+}
+
+/**
+ * Requests to execute this webhook.
+ *
+ * This is a 'fire and forget' variant of [execute]. It will not wait for a response and might not throw an
+ * Exception if the request wasn't executed.
+ */
+suspend inline fun WebhookBehavior.executeIgnored(token: String, builder: ExecuteWebhookBuilder.() -> Unit) {
+    kord.rest.webhook.executeWebhook(token = token, webhookId = id.value, wait = false, builder = builder)
 }